requirements.txt
asgiref==3.5.2
Django==3.2
django-smart-selects==1.5.9
djangorestframework==3.13.1
Pillow==9.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
drf-yasg==1.20.0
django-filter==22.1